Some test suite updates.
Add tst-key4 from libc, although we need to deal with 1 less pthread key
than normal due to something that's going on with pthread when libeatmydata
is present that I don't quite know what's going on.

We also update tst-cancel and the test run script so that we better capture
errors from the test suite.

+++ test_run.sh
@@ -55,6 +55,11 @@ DYLD_LIBRARY_PATH=libeatmydata/.libs
 export DYLD_LIBRARY_PATH
 
 LD_PRELOAD=./.libs/libeatmydata.so DYLD_FORCE_FLAT_NAMESPACE=1 DYLD_INSERT_LIBRARIES=./.libs/libeatmydata.dylib strace -o test.result.run $1 $LIBEATMYDATA_TEST_ARGS
+eret=$?
+if [ $eret != 0 ]; then
+	exit 2;
+fi
+
 grep '^[a-z]*sync\|O_SYNC' test.result.run
 ret=$?
 rm test.result.run
